Add files via upload
This project is for a most simple hostel management system, which we can use it in many platforms. the main feature of this code is that it is easy to use for every individual and if we want to add a new feature, we can add it with less cost.